+++ pkg/local_object_storage/shard/control_test.go
@@ -122,9 +122,6 @@ func TestResyncMetabase(t *testing.T) {
 
 	defer os.RemoveAll(p)
 
-	testObj := objecttest.Object()
-	writeCacheThreshold := len(testObj.Marshal())
-
 	shID := ID("test")
 	sh := New(
 		WithBlobstor(fstree.New(
@@ -140,7 +137,6 @@ func TestResyncMetabase(t *testing.T) {
 		WithWriteCache(true),
 		WithWriteCacheOptions(
 			writecache.WithPath(filepath.Join(p, "wc")),
-			writecache.WithMaxObjectSize(uint64(writeCacheThreshold)),
 			writecache.WithLogger(zaptest.NewLogger(t)),
 		),
 	)
@@ -164,8 +160,7 @@ func TestResyncMetabase(t *testing.T) {
 		obj.SetType(objectSDK.TypeRegular)
 
 		if i < objNum/2 {
-			// this object goes to blobstor directly for sure
-			payload := make([]byte, writeCacheThreshold)
+			payload := make([]byte, 1024)
 			_, err := rand.Read(payload)
 			require.NoError(t, err)
 
@@ -307,7 +302,6 @@ func TestResyncMetabase(t *testing.T) {
 		WithWriteCache(true),
 		WithWriteCacheOptions(
 			writecache.WithPath(filepath.Join(p, "wc")),
-			writecache.WithMaxObjectSize(uint64(writeCacheThreshold)),
 			writecache.WithLogger(zaptest.NewLogger(t)),
 		),
 	)